py_bind/requirements.txt
cython>=3.0.0
nose
numpy>=1.21.0
pyyaml
